weather_lcd/requirements.txt
pyserial
requests

url = "freegeoip.net/json/{IP}"
